requirements.txt
lightgbm==4.4.0
matplotlib==3.7.5
networkx==3.1
numba==0.58.1
numpy==1.24.4
ogb==1.3.6
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.3.2
scipy==1.14.0
torch==2.2.1
torch_geometric==2.5.3
torch_scatter==2.1.2+pt22cu118
torch_sparse==0.6.18+pt22cu118
tqdm==4.66.4
